+++ app/views/admin/center/center/index.erb
@@ -1,2 +1,1 @@
 <h1>管理員區域</h1>
-<%= link_to '管理員使用者區域', admin_center_users_path %>
