+++ app/views/users/show.erb
@@ -1,0 +1,11 @@
+<h1>檢視</h1>
+<table>
+  <tr>
+    <th>name</th>
+    <th>email</th>
+  </tr>
+  <tr>
+    <td><%= @user.name %></td>
+    <td><%= @user.email %></td>
+  </tr>
+</table>
