+++ app/views/welcome/index.erb
@@ -1,2 +1,3 @@
 <h1>這是首頁</h1>
 <%= link_to '前往管理員區域', admin_center_index_path %>
+<%= link_to '註冊會員', new_user_path %>
